Fix segment world width/height computation
Previously, this was computed using the ratio between the (pixel
center and the camera center distance) and the (world point value
and origin distance) - the problem here is that if the object was
near the origin along either x or y axis, the ratio became ill-
conditioned.

Now, this is computed using the ratio between focal distance and
object depth.

Signed-off-by: Mason Peterson <mbpeterson70@gmail.com>