Make submap creation dependent on time
submap_align.py orinally only looked to see if the center of an object
was within a distance threshold of the submap center - now it
additionally checks to make sure it was seen around the same time. This
prevents duplicates of objects when a submap area was visited multiple
times by one robot.

Signed-off-by: Mason Peterson <mbpeterson70@gmail.com>